backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.8.0
qdrant-client>=1.7.1
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
cohere>=4.0.0